Convert iterator test to unit test
